# container/ring: Fix various data races and race conditions