--- a/pkg/container/ring.go
+++ b/pkg/container/ring.go
@@ -42,10 +42,11 @@ type Ring struct {
 	dataLen uint64
 	// data is the internal buffer of this ring buffer.
 	data []*v1.Event
-	// cond is used to signal when a write was made so a "waiting" reader in
-	// ReadFrom(<-chan struct{}, uint64) <-chan *pb.Payload knows the writer
-	// has written into the internal buffer.
-	cond *sync.Cond
+	// notify{Mu,Ch} are used to signal a waiting reader in readFrom when the
+	// writer has written a new value.
+	// We cannot use sync.Cond as it cannot be used in select statements.
+	notifyMu sync.Mutex
+	notifyCh chan struct{}
 }
 
 // NewRing creates a ring buffer. For efficiency, the internal
@@ -71,7 +72,8 @@ func NewRing(n int) *Ring {
 		cycleMask: ^uint64(0) >> cycleExp,
 		dataLen:   dataLen,
 		data:      make([]*v1.Event, dataLen, dataLen),
-		cond:      sync.NewCond(&sync.RWMutex{}),
+		notifyMu:  sync.Mutex{},
+		notifyCh:  nil,
 	}
 }
 
@@ -111,10 +113,27 @@ func (r *Ring) Cap() uint64 {
 // writing block. The entry must not be nil, otherwise readFrom will block when
 // reading back this event.
 func (r *Ring) Write(entry *v1.Event) {
+	// We need to lock the notification mutex when updating r.write, otherwise
+	// there is a race condition where a readFrom goroutine goes to sleep
+	// after we sent out the notification.
+	// This lock is only shared with other readFrom instances that are about
+	// to go to sleep, so contention should be low. Notably, readers which are
+	// far away from the current write pointer will still be able to make
+	// progress concurrently.
+
+	r.notifyMu.Lock()
+
 	write := atomic.AddUint64(&r.write, 1)
 	writeIdx := (write - 1) & r.mask
 	r.dataStoreAtomic(writeIdx, entry)
-	r.cond.Broadcast()
+
+	// notify any sleeping readers
+	if r.notifyCh != nil {
+		close(r.notifyCh)
+		r.notifyCh = nil
+	}
+
+	r.notifyMu.Unlock()
 }
 
 // LastWriteParallel returns the last element written.
@@ -179,14 +198,11 @@ func (r *Ring) readFrom(ctx context.Context, read uint64) <-chan *v1.Event {
 		// a half cycle is (^uint64(0)/r.dataLen)/2
 		// which translates into (^uint64(0)>>r.dataLen)>>1
 		halfCycle := (^uint64(0) >> r.cycleExp) >> 1
-		go func() {
-			<-ctx.Done()
-			r.cond.Broadcast()
-		}()
 		defer func() {
 			close(ch)
 		}()
-		// read forever until stop is closed
+
+		// read forever until ctx is done
 		for ; ; read++ {
 			readIdx := read & r.mask
 			event := r.dataLoadAtomic(readIdx)
@@ -247,34 +263,35 @@ func (r *Ring) readFrom(ctx context.Context, read uint64) <-chan *v1.Event {
 			case event == nil || readCycle >= (writeCycle+1)&r.cycleMask && readCycle < (halfCycle+writeCycle)&r.cycleMask:
 				// The writer has already written a new event so there's no
 				// need to stop the reader.
-				//
-				// FIXME?: It is still possible the writer will
-				//  write something after we check for `r.write` and before
-				//  we do r.cond.Wait, which means we will only receive
-				//  a broadcast() from the writer after a new event arrives,
-				//  making the reader block.
-				//  We can sort of avoiding it by checking if r.write got
-				//  incremented before we block while we wait for a broadcast
+
+				// Before going to sleep, we need to check that there has been
+				// no write in the meantime. This check can only be race free
+				// if the lock on notifyMu is held, otherwise a write can occur
+				// before we obtain the notifyCh instance.
+				r.notifyMu.Lock()
 				if lastWrite != atomic.LoadUint64(&r.write)-1 {
+					// A write has occurred - retry
+					r.notifyMu.Unlock()
 					read--
 					continue
 				}
-				r.cond.L.Lock()
+
+				// This channel will be closed by the writer if it makes a write
+				if r.notifyCh == nil {
+					r.notifyCh = make(chan struct{})
+				}
+				notifyCh := r.notifyCh
+				r.notifyMu.Unlock()
+
+				// Sleep until a write occurs or the context is cancelled
 				select {
-				case <-ctx.Done():
-					r.cond.L.Unlock()
-					return
-				default:
-					r.cond.Wait()
-					r.cond.L.Unlock()
+				case <-notifyCh:
 					read--
-				}
-				select {
+					continue
 				case <-ctx.Done():
 					return
-				default:
-					continue
 				}
+
 			}
 		}
 	}()