This is a first step towards fixing #82. This uses
`atomic.{Load,Store}Pointer` to access the individual slots in the ring
buffer, ensuring that each write is properly propagated to all other
threads. While it seems to satisfy the Go race detector, it seems that
running the unit test now with `-race` will now sometimes fail with a
semantic error, i.e. the ring buffer content is not what is expected.
This indicates that there remains a real race condition somewhere, this
commit only addresses the data race discovered by the Go race detector.

This means that #142 remains blocked, as we also have to find the root
cause for the race condition. Tests pass consistently when running
this commit without `-race` however, therefore it does not seem to
introduce any regressions.

Signed-off-by: Sebastian Wicki <sebastian@isovalent.com>