@gandro 2020-03-05
Performance impact on my laptop:

Before:
```
goos: linux
goarch: amd64
pkg: github.com/cilium/hubble/pkg/container
BenchmarkRingWrite-8   	116604054	        10.1 ns/op	       0 B/op	       0 allocs/op
BenchmarkRingRead-8    	160090080	        12.8 ns/op	       0 B/op	       0 allocs/op
PASS
ok  	github.com/cilium/hubble/pkg/container	10.292s

```

After:
```
goos: linux
goarch: amd64
pkg: github.com/cilium/hubble/pkg/container
BenchmarkRingWrite-8   	72918658	        16.4 ns/op	       0 B/op	       0 allocs/op
BenchmarkRingRead-8    	287611406	        18.1 ns/op	       0 B/op	       0 allocs/op
PASS
ok  	github.com/cilium/hubble/pkg/container	22.051s
```