@gandro 2020-03-05
Seems like the failed unit test that I discovered is a flake/timeout which only occurs when `-race` is enabled. I guess that's expected, since the Go race detector can impose quite the overhead. See https://github.com/cilium/hubble/pull/177/commits/35ff424ec9bd499977c0ca9a9c36a78d77f0ec22

This means that with this PR we likely can merge #142  and close #82.