@Rolinh 2020-03-05
I've just hit a ring reader test failure with these changes:
```
--- FAIL: TestRingReader_NextFollow (0.32s)
    --- FAIL: TestRingReader_NextFollow/read_5,_start_at_position_5,_expect_timeout=false (0.22s)
        ring_reader_test.go:219: 
            	Error Trace:	ring_reader_test.go:219
            	Error:      	Not equal: 
            	            	expected: []*v1.Event{(*v1.Event)(0xc000537880), (*v1.Event)(0xc0005378a0), (*v1.Event)(0xc0005378c0), (*v1.Event)(0xc0005378e0), (*v1.Event)(0xc000537920)}
            	            	actual  : []*v1.Event{(*v1.Event)(0xc000537640), (*v1.Event)(0xc000537660), (*v1.Event)(0xc0005376a0), (*v1.Event)(nil), (*v1.Event)(nil)}
            	            	
            	            	Diff:
            	            	--- Expected
            	            	+++ Actual
            	            	@@ -13,10 +13,4 @@
            	            	  }),
            	            	- (*v1.Event)({
            	            	-  Timestamp: (*types.Timestamp)(1970-01-01T00:00:08Z),
            	            	-  Event: (interface {}) <nil>
            	            	- }),
            	            	- (*v1.Event)({
            	            	-  Timestamp: (*types.Timestamp)(1970-01-01T00:00:09Z),
            	            	-  Event: (interface {}) <nil>
            	            	- })
            	            	+ (*v1.Event)(<nil>),
            	            	+ (*v1.Event)(<nil>)
            	            	 }
            	Test:       	TestRingReader_NextFollow/read_5,_start_at_position_5,_expect_timeout=false
        ring_reader_test.go:220: 
            	Error Trace:	ring_reader_test.go:220
            	Error:      	Not equal: 
            	            	expected: false
            	            	actual  : true
            	Test:       	TestRingReader_NextFollow/read_5,_start_at_position_5,_expect_timeout=false
FAIL
```