@gandro 2020-03-11
I have identified a few race conditions and believe to have fixed them. See commit messages for detail. I have not observed any data race anymore, so I'm confident that this fixes #82.